aergocli: use relative path when processing import statements
